+++ source_gen/test/builder_test.dart
@@ -21,12 +21,12 @@ void main() {
 
   test('Bad generated source', () async {
     var srcs = _createPackageStub();
-    var builder = new PartBuilder([const _BadOutputGenerator()]);
+    var builder = new PartBuilder([const _BadOutputGenerator()], '.foo.dart');
 
     await testBuilder(builder, srcs,
         generateFor: new Set.from(['$_pkgName|lib/test_lib.dart']),
         outputs: {
-          '$_pkgName|lib/test_lib.g.dart':
+          '$_pkgName|lib/test_lib.foo.dart':
               decodedMatches(contains('not valid code!')),
         });
   });
@@ -67,29 +67,30 @@ void main() {
       () async {
     expect(
         () => new PartBuilder(
-            [const CommentGenerator(), const _LiteralGenerator()]),
+            [const CommentGenerator(), const _LiteralGenerator()], '.foo.dart'),
         returnsNormally);
   });
 
   test('Allow no "library"  by default', () async {
     var sources = _createPackageStub(testLibContent: 'class A {}');
-    var builder = new PartBuilder([const CommentGenerator()]);
+    var builder = new PartBuilder([const CommentGenerator()], '.foo.dart');
 
     await testBuilder(builder, sources,
-        outputs: {'$_pkgName|lib/test_lib.g.dart': _testGenNoLibrary});
+        outputs: {'$_pkgName|lib/test_lib.foo.dart': _testGenNoLibrary});
   });
 
   test('Does not fail when there is no output', () async {
     var sources = _createPackageStub(testLibContent: 'class A {}');
-    var builder = new PartBuilder([const CommentGenerator(forClasses: false)]);
+    var builder = new PartBuilder(
+        [const CommentGenerator(forClasses: false)], '.foo.dart');
     await testBuilder(builder, sources, outputs: {});
   });
 
   test('Use new part syntax when no library directive exists', () async {
     var sources = _createPackageStub(testLibContent: 'class A {}');
-    var builder = new PartBuilder([const CommentGenerator()]);
+    var builder = new PartBuilder([const CommentGenerator()], '.foo.dart');
     await testBuilder(builder, sources,
-        outputs: {'$_pkgName|lib/test_lib.g.dart': _testGenNoLibrary});
+        outputs: {'$_pkgName|lib/test_lib.foo.dart': _testGenNoLibrary});
   });
 
   test(
@@ -112,17 +113,17 @@ void main() {
 
   test('handle generator errors well', () async {
     var srcs = _createPackageStub(testLibContent: _testLibContentWithError);
-    var builder = new PartBuilder([const CommentGenerator()]);
+    var builder = new PartBuilder([const CommentGenerator()], '.foo.dart');
     await testBuilder(builder, srcs,
         generateFor: new Set.from(['$_pkgName|lib/test_lib.dart']),
         outputs: {
-          '$_pkgName|lib/test_lib.g.dart': _testGenPartContentError,
+          '$_pkgName|lib/test_lib.foo.dart': _testGenPartContentError,
         });
   });
 
   test('warns when a non-standalone builder does not see "part"', () async {
     var srcs = _createPackageStub(testLibContent: _testLibContentNoPart);
-    var builder = new PartBuilder([const CommentGenerator()]);
+    var builder = new PartBuilder([const CommentGenerator()], '.foo.dart');
     var logs = <String>[];
     await testBuilder(
       builder,
@@ -131,7 +132,7 @@ void main() {
         logs.add(log.message);
       },
     );
-    expect(logs, ['Missing "part \'test_lib.g.dart\';".']);
+    expect(logs, ['Missing "part \'test_lib.foo.dart\';".']);
   });
 
   test('generator with an empty result creates no outputs', () async {
@@ -161,41 +162,43 @@ void main() {
       builder,
       srcs,
       outputs: {
-        '$_pkgName|lib/test_lib.g.dart': _whitespaceTrimmed,
+        '$_pkgName|lib/test_lib.foo.dart': _whitespaceTrimmed,
       },
     );
   });
 
   test('defaults to formatting generated code with the DartFormatter',
       () async {
-    await testBuilder(new PartBuilder([const UnformattedCodeGenerator()]),
+    await testBuilder(
+        new PartBuilder([const UnformattedCodeGenerator()], '.foo.dart'),
         {'$_pkgName|lib/a.dart': 'library a; part "a.part.dart";'},
         generateFor: new Set.from(['$_pkgName|lib/a.dart']),
         outputs: {
-          '$_pkgName|lib/a.g.dart':
+          '$_pkgName|lib/a.foo.dart':
               decodedMatches(contains(UnformattedCodeGenerator.formattedCode)),
         });
   });
 
   test('PartBuilder uses a custom header when provided', () async {
     await testBuilder(
-        new PartBuilder([const UnformattedCodeGenerator()],
+        new PartBuilder([const UnformattedCodeGenerator()], '.foo.dart',
             header: _customHeader),
         {'$_pkgName|lib/a.dart': 'library a; part "a.part.dart";'},
         generateFor: new Set.from(['$_pkgName|lib/a.dart']),
         outputs: {
-          '$_pkgName|lib/a.g.dart':
+          '$_pkgName|lib/a.foo.dart':
               decodedMatches(startsWith('$_customHeader\n\npart of')),
         });
   });
 
   test('PartBuilder includes no header when `header` is empty', () async {
     await testBuilder(
-        new PartBuilder([const UnformattedCodeGenerator()], header: ''),
+        new PartBuilder([const UnformattedCodeGenerator()], '.foo.dart',
+            header: ''),
         {'$_pkgName|lib/a.dart': 'library a; part "a.part.dart";'},
         generateFor: new Set.from(['$_pkgName|lib/a.dart']),
         outputs: {
-          '$_pkgName|lib/a.g.dart': decodedMatches(startsWith('part of')),
+          '$_pkgName|lib/a.foo.dart': decodedMatches(startsWith('part of')),
         });
   });
 
@@ -290,12 +293,12 @@ void main() {
 
   test('can skip formatting with a trivial lambda', () async {
     await testBuilder(
-        new PartBuilder([const UnformattedCodeGenerator()],
+        new PartBuilder([const UnformattedCodeGenerator()], '.foo.dart',
             formatOutput: (s) => s),
         {'$_pkgName|lib/a.dart': 'library a; part "a.part.dart";'},
         generateFor: new Set.from(['$_pkgName|lib/a.dart']),
         outputs: {
-          '$_pkgName|lib/a.g.dart': decodedMatches(
+          '$_pkgName|lib/a.foo.dart': decodedMatches(
               contains(UnformattedCodeGenerator.unformattedCode)),
         });
   });
@@ -303,12 +306,12 @@ void main() {
   test('can pass a custom formatter with formatOutput', () async {
     var customOutput = 'final String hello = "hello";';
     await testBuilder(
-        new PartBuilder([const UnformattedCodeGenerator()],
+        new PartBuilder([const UnformattedCodeGenerator()], '.foo.dart',
             formatOutput: (_) => customOutput),
         {'$_pkgName|lib/a.dart': 'library a; part "a.part.dart";'},
         generateFor: new Set.from(['$_pkgName|lib/a.dart']),
         outputs: {
-          '$_pkgName|lib/a.g.dart': decodedMatches(contains(customOutput)),
+          '$_pkgName|lib/a.foo.dart': decodedMatches(contains(customOutput)),
         });
   });
 
@@ -319,20 +322,20 @@ void main() {
     final builders = new PartBuilder([
       const _LiteralGenerator(),
       const _LiteralGenerator(),
-    ]);
+    ], '.foo.dart');
     expect(builders.toString(),
-        'Generating .g.dart: _LiteralGenerator, _LiteralGenerator');
+        'Generating .foo.dart: _LiteralGenerator, _LiteralGenerator');
   });
 }
 
 Future _generateTest(CommentGenerator gen, String expectedContent) async {
   var srcs = _createPackageStub();
-  var builder = new PartBuilder([gen]);
+  var builder = new PartBuilder([gen], '.foo.dart');
 
   await testBuilder(builder, srcs,
       generateFor: new Set.from(['$_pkgName|lib/test_lib.dart']),
       outputs: {
-        '$_pkgName|lib/test_lib.g.dart': decodedMatches(expectedContent),
+        '$_pkgName|lib/test_lib.foo.dart': decodedMatches(expectedContent),
       },
       onLog: (log) => fail('Unexpected log message: ${log.message}'));
 }
@@ -341,12 +344,14 @@ Map<String, String> _createPackageStub(
     {String testLibContent, String testLibPartContent}) {
   return {
     '$_pkgName|lib/test_lib.dart': testLibContent ?? _testLibContent,
-    '$_pkgName|lib/test_lib.g.dart': testLibPartContent ?? _testLibPartContent,
+    '$_pkgName|lib/test_lib.foo.dart':
+        testLibPartContent ?? _testLibPartContent,
   };
 }
 
 PartBuilder _unformattedLiteral([String content]) =>
-    new PartBuilder([new _LiteralGenerator(content)], formatOutput: (s) => s);
+    new PartBuilder([new _LiteralGenerator(content)], '.foo.dart',
+        formatOutput: (s) => s);
 
 /// Returns the [String] provided in the constructor, or `null`.
 class _LiteralGenerator extends Generator {
@@ -371,7 +376,7 @@ const _pkgName = 'pkg';
 
 const _testLibContent = r'''
 library test_lib;
-part 'test_lib.g.dart';
+part 'test_lib.foo.dart';
 final int foo = 42;
 class Person { }
 ''';